backend/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonpickle==2.0.0
kombu==5.1.0
MarkupSafe==2.0.1
packaging==21.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
PyJWT==2.1.0
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.13
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
